rename discordmoviesattributes.py to attributes.py
